Suggested adjustments are made
