+++ src/test/java/runners/RunCukesTest.java
@@ -2,10 +2,7 @@ package runners;
 
 import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.runner.RunWith;
-import utils.BrowserManagement;
 
 
 @RunWith(Cucumber.class)
@@ -14,17 +11,6 @@ import utils.BrowserManagement;
         features = "src/test/resources/features/Sample.feature",
         glue = {"steps"}
 )
-class RunCukesTest {
+class RunCukesTest extends AbstractTest {
 
-    @BeforeClass
-    public static void init() {
-        BrowserManagement.startServer();
-    }
-
-    @AfterClass
-    public static void finish() {
-        BrowserManagement.getDriver().closeApp();
-        BrowserManagement.getDriver().quit();
-        BrowserManagement.stopServer();
-    }
 }
